Create tables with sample row in each
